Fixed issue: Dragging card and going out of bound
Fixed the issued with dragging the card only working when clicking at a
rectangular area based at point (0,0). The card also can no longer move
outside the bounds of the screen. I created 2 debugger lines that spit
information about the mouse, and the card in the debug output.  We can
now move on to implement tableaus and identifying multiple cards.